sudoku-web/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.5.0
waitress==2.2.0
pytest==7.4.0
ruff==0.28.0